Create Bowl
This is a "random number generator" in the style of a bowl filled with tickets. In reality, you would think of a number that you'd like to find, draw a ticket with a number from the bowl, if the number is not what you thought of, then you would put the ticket back in the bowl. It logs the amount of tries, and the numbers drawn. Drawing the same numbers is a part of program.  Further explanation below. 

The number you want to find, is the highest number it will guess. 
So if you guess for "1000 000 (one million)" prepare for a long wait until it finds the number. As mentioned earlier, the numbers can and probably will repeat. 
This function was added because if you set a specified number, it should be as high as the guess, but if the guess is 4, out of 1 million, it will take forever to find. Therefore this function prevents that from happening, if you want to find 4, the program will only search between the range 0-4.